clean up creator project namespaces
